Implement Acces Guard for replies, votes and threads routes

+++ app/api/vote.py
@@ -17,12 +17,16 @@ class Vote(BaseModel):
 @router.get("/votes", tags=["vote"])
 async def get_votes(request: Request):
     try:
+        user = request.user
         params = request.query_params
-        if not params:
+        query_dict = dict(params)
+
+        if user and user.get("role") != 'ADMIN':
+            query_dict["userEmail"] = user.get('email')
+
+        if not query_dict:
             votes = await prisma.vote.find_many()
         else:
-            query_dict = dict(params)
-
             if(params.get("id")):
                 query_dict["id"] = int(query_dict["id"])
 
@@ -45,13 +49,21 @@ async def get_votes(request: Request):
         raise HTTPException(status_code=404, detail="Votes not found")
 
 @router.get("/votes/{vote_id}", tags=["vote"])
-async def get_vote(vote_id : int):
-    vote = await prisma.vote.find_unique(where={'id':vote_id })
+async def get_vote(vote_id : int, request: Request):
+    try:
+        user = request.user
+        vote = await prisma.vote.find_unique(where={'id':vote_id })
 
-    if not vote:
-        raise HTTPException(status_code=404, detail="Vote not found")
-    
-    return vote
+        if not vote:
+            raise HTTPException(status_code=404, detail="Vote not found")
+        
+        if user.get('email') != vote.userEmail and user.get('role') != 'ADMIN':
+            raise HTTPException(status_code=403, detail="Vote belong to another user")      
+
+        return vote
+    except Exception as e:
+        print(e)
+        raise HTTPException(status_code=404, detail=e.detail)
 
 @router.post("/votes", tags=["vote"])
 async def add_vote(body: Vote):
@@ -72,13 +84,18 @@ async def add_vote(body: Vote):
         raise HTTPException(status_code=404, detail="Failed to post vote")
 
 @router.patch("/votes/{vote_id}", tags=["vote"])
-async def update_vote(vote_id:int, body: dict):
+async def update_vote(vote_id:int, body: dict, request: Request):
     try:
+        user = request.user
         vote = await prisma.vote.find_unique(where={'id': vote_id })
+        print(type(vote))
 
         if not vote:
             raise HTTPException(status_code=404, detail="Vote not found")
         
+        if user.get('email') != vote.userEmail and user.get('role') != 'ADMIN':
+            raise HTTPException(status_code=403, detail="Vote belong to another user")    
+
         vote = await prisma.vote.update(
             where= { 'id': vote_id },
             data= dict(body)
@@ -86,16 +103,20 @@ async def update_vote(vote_id:int, body: dict):
 
         return vote
     except Exception as e:
-        print(e)
-        raise HTTPException(status_code=404, detail="Failed to patch vote")
+        print(e)    
+        raise HTTPException(status_code=404, detail=e.detail)
 
 @router.delete("/votes/{vote_id}", tags=["vote"])
-async def update_vote(vote_id:int):
+async def update_vote(vote_id:int, request: Request):
     try:
+        user = request.user
         vote = await prisma.vote.find_unique(where={'id': vote_id })
 
         if not vote:
             raise HTTPException(status_code=404, detail="Vote not found")
+        
+        if user.get('email') != vote.userEmail and user.get('role') != 'ADMIN':
+            raise HTTPException(status_code=403, detail="Vote belong to another user")      
         
         vote = await prisma.vote.delete(
             where= { 'id': vote_id }
@@ -104,4 +125,4 @@ async def update_vote(vote_id:int):
         return vote
     except Exception as e:
         print(e)
-        raise HTTPException(status_code=404, detail="Failed to delete vote")
+        raise HTTPException(status_code=404, detail=e.detail)
